chore: move metadata types
