+++ src/userscripts/entries/userscript-metadata.ts
